extract classifyterminator to shared utility, eliminating three-way sync requirement
