+++ src/codegen/infrastructure/base-generator.ts
@@ -1,4 +1,5 @@
 import { Expression } from "../../ast/types.js";
+import { classifyTerminator } from "./terminator-classifier.js";
 import {
   SymbolTable,
   SymbolKind,
@@ -468,14 +469,7 @@ export class BaseGenerator {
   }
 
   protected classifyTerminator(instruction: string): boolean {
-    const trimmed = instruction.trim();
-    return (
-      trimmed.startsWith("ret ") ||
-      trimmed === "ret void" ||
-      trimmed.startsWith("br ") ||
-      trimmed.startsWith("unreachable") ||
-      trimmed.startsWith("switch ")
-    );
+    return classifyTerminator(instruction);
   }
 
   lastInstructionIsTerminator(): boolean {
